+++ domestic/models.py
@@ -14,7 +14,11 @@ from wagtail.images import get_image_model_string
 from wagtail.images.edit_handlers import ImageChooserPanel
 
 from core import blocks as core_blocks, cms_slugs, forms, helpers, mixins
-from core.constants import ARTICLE_TYPES, VIDEO_TRANSCRIPT_HELP_TEXT
+from core.constants import (
+    ARTICLE_TYPES,
+    RICHTEXT_FEATURES__REDUCED,
+    VIDEO_TRANSCRIPT_HELP_TEXT,
+)
 from core.helpers import build_social_links
 from core.models import CMSGenericPage, Country, IndustryTag, Region, Tag
 from directory_constants import choices
@@ -81,7 +85,11 @@ class DomesticHomePage(
     Page,
 ):
 
-    body = RichTextField(null=True, blank=True)
+    body = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+        null=True,
+        blank=True,
+    )
     button = StreamField([('button', core_blocks.ButtonBlock(icon='cog'))], null=True, blank=True)
     image = models.ForeignKey(
         get_image_model_string(), null=True, blank=True, on_delete=models.SET_NULL, related_name='+'
@@ -166,7 +174,7 @@ class TopicLandingBasePage(BaseContentPage):
         null=True,
         blank=True,
     )
-    banner_text = RichTextField(blank=True)
+    banner_text = RichTextField(features=RICHTEXT_FEATURES__REDUCED, blank=True)
     teaser = models.TextField(blank=True)
 
 
@@ -389,6 +397,7 @@ class CountryGuidePage(cms_panels.CountryGuidePagePanels, BaseContentPage):
     )
 
     section_one_body = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
         null=True,
         verbose_name='3 unique selling points markdown',
         help_text='Use H2s for the 3 subheadings',
@@ -470,6 +479,7 @@ class CountryGuidePage(cms_panels.CountryGuidePagePanels, BaseContentPage):
         verbose_name="Summary for 'Tax and customs'",
     )
     fact_sheet_column_1_body = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
         blank=True,
         verbose_name="Detailed text for 'Tax and customs'",
         help_text='Use H4 for each subcategory heading. Maximum five sub categories. Aim for 50 words each.',
@@ -485,6 +495,7 @@ class CountryGuidePage(cms_panels.CountryGuidePagePanels, BaseContentPage):
         verbose_name="Summary for 'Protecting your business'",
     )
     fact_sheet_column_2_body = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
         blank=True,
         verbose_name="Detailed text for 'Protecting your business'",
         help_text='Use H4 for each sub category heading. Maximum five sub categories. Aim for 50 words each.',
@@ -647,11 +658,14 @@ class ArticlePage(
         related_name='+',
     )
     article_video_transcript = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
         null=True,
         blank=True,
         help_text=VIDEO_TRANSCRIPT_HELP_TEXT,
     )
-    article_body_text = RichTextField()
+    article_body_text = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+    )
 
     cta_title = models.CharField(
         max_length=255,
@@ -740,6 +754,7 @@ class ArticleListingPage(cms_panels.ArticleListingPagePanels, BaseContentPage):
     )
 
     list_teaser = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
         null=True,
         blank=True,
     )
@@ -768,7 +783,9 @@ class CampaignPage(cms_panels.CampaignPagePanels, BaseContentPage):
     )
 
     section_one_heading = models.CharField(max_length=255)
-    section_one_intro = RichTextField()
+    section_one_intro = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+    )
     section_one_image = models.ForeignKey(
         'core.AltTextImage',
         null=True,
@@ -785,7 +802,9 @@ class CampaignPage(cms_panels.CampaignPagePanels, BaseContentPage):
         related_name='+',
     )
     selling_point_one_heading = models.CharField(max_length=255)
-    selling_point_one_content = RichTextField()
+    selling_point_one_content = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+    )
 
     selling_point_two_icon = models.ForeignKey(
         'core.AltTextImage',
@@ -799,7 +818,11 @@ class CampaignPage(cms_panels.CampaignPagePanels, BaseContentPage):
         null=True,
         blank=True,
     )
-    selling_point_two_content = RichTextField(null=True, blank=True)
+    selling_point_two_content = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+        null=True,
+        blank=True,
+    )
 
     selling_point_three_icon = models.ForeignKey(
         'core.AltTextImage',
@@ -814,6 +837,7 @@ class CampaignPage(cms_panels.CampaignPagePanels, BaseContentPage):
         blank=True,
     )
     selling_point_three_content = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
         null=True,
         blank=True,
     )
@@ -832,7 +856,9 @@ class CampaignPage(cms_panels.CampaignPagePanels, BaseContentPage):
     section_two_heading = models.CharField(
         max_length=255,
     )
-    section_two_intro = RichTextField()
+    section_two_intro = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+    )
 
     section_two_image = models.ForeignKey(
         'core.AltTextImage',
@@ -856,7 +882,9 @@ class CampaignPage(cms_panels.CampaignPagePanels, BaseContentPage):
     related_content_heading = models.CharField(
         max_length=255,
     )
-    related_content_intro = RichTextField()
+    related_content_intro = RichTextField(
+        features=RICHTEXT_FEATURES__REDUCED,
+    )
 
     related_page_one = models.ForeignKey(
         'domestic.ArticlePage',